Make all commands in CLI tests system independent
Replaces POSIX-specific commands and path names in the RSPEC CLI tests
with Ruby utility commands that work on all systems. Commands
like 'mkdir' and 'cd' get replaced with FileUtils.mkdir_p() and
Dir.chdir() respectively. Pathnames are formed with File.join(),
allowing for valid comparisons on all systems. Also, 'run_system_comand'
is introduced as a replacement for system() to capture errors

Conversions are added to fix the parsing of OXEN_API_KEY in
spech_helper, and a canonicalization inconsistency in schemas_add is also fixed