Check modified time before file hash in checkout
Implemented a new utility function to compare a file in the working directory
against its node in the merkle tree. It first checks the file's modified time,
and if that differs from the node, only then is the file hashed for comparison.
This avoids the unnecessary re-hashing of markedly unmodified files in the checkout code