Optimize checkout logic with seen files hashset
Speeds up r_remove_if_not_in_target in the checkout logic by
implementing a hashset of seen files from the target tree. Previously,
that function would have to lookup every file it was checking for
removal in the Merkle tree. To avoid this, we now do
r_restore_missing_or_modified_files first, and store every file it comes
across in a hashset, so that the check for its presence in the from tree
is immediate